created a data dot sql file and organized a package for security so we can test with the created data as well
