--- src/main/java/nl/garagemeijer/salesapi/models/Security.java
@@ -1,8 +1,0 @@
-package nl.garagemeijer.salesapi.models;
-
-
-public class Security {
-
-    //    authorisatie
-    private String userType; // admin of salesperson
-}
